Got some help but still figuring it out.
Arrays half implemented.
Randomise values implemented

+++ Assets/22T3_GAD170.2_Bundalian_John/Plant.cs
@@ -8,7 +8,7 @@ namespace JohnBundalian
     {
         private string name;
         private Herbs species = Herbs.Undefined;
-        private int healthiness;
+        private float healthiness;
         private int price;
 
         void Start()
@@ -18,18 +18,19 @@ namespace JohnBundalian
             myPlant = Herbs.BlackLotus;
         }
         //Variables:
-        public int GetHealthiness()
+        public float GetHealthiness()
         {
+            
             return healthiness;
         }
 
-        public void SetInitialStats(string newName, Herbs newHerbs, int newHealthiness, int newPrice)
+        public void SetInitialStats(string newName, Herbs newHerbs, float newHealthiness, int newPrice)
         {
             name = newName;
             species = newHerbs;
             healthiness = newHealthiness;
             price = newPrice;
-
+            
            // newHerbs.herbsName = Herbs[UnityEngine.Random.Range(0, HerbsNames.Count())];
              //   newHealthiness.maxHealth = Random.Range(1, 10);
            // newPrice.maxPrice = Random.Range(1, 20);
